refactor modal into a new component
