+++ site/src/app/roadmap/planner/PlannerLoader.tsx
@@ -1,6 +1,5 @@
 'use client';
 import { FC, useCallback, useEffect, useState } from 'react';
-import { Button, Dialog, DialogContent, DialogContentText, DialogTitle, Divider } from '@mui/material';
 import {
   collapseAllPlanners,
   expandAllPlanners,
@@ -36,6 +35,7 @@ import { setDataLoadState } from '../../../store/slices/transferCreditsSlice';
 import { compareRoadmaps, restoreRevision } from '../../../helpers/roadmap';
 import { deepCopy } from '../../../helpers/util';
 import { setCustomCourses } from '../../../store/slices/customCourseSlice';
+import PlannerLoaderModal from './PlannerLoaderModal';
 
 function useCheckUnsavedChanges() {
   const currentIndex = useAppSelector((state) => state.roadmap.currentRevisionIndex);
@@ -231,126 +231,17 @@ const PlannerLoader: FC = () => {
     setShowSyncModal(false);
   };
 
-  const getRelativeTime = (timestamp: string | undefined) => {
-    if (!timestamp) return undefined;
-    const diff = new Date(timestamp).getTime() - Date.now();
-    const seconds = Math.round(diff / 1000);
-    const minutes = Math.round(seconds / 60);
-    const hours = Math.round(minutes / 60);
-    const days = Math.round(hours / 24);
-
-    const rtf = new Intl.RelativeTimeFormat('en', { numeric: 'auto' });
-
-    if (Math.abs(seconds) < 60) return rtf.format(seconds, 'second');
-    if (Math.abs(minutes) < 60) return rtf.format(minutes, 'minute');
-    if (Math.abs(hours) < 24) return rtf.format(hours, 'hour');
-    return rtf.format(days, 'day');
-  };
-
-  const countRoadmapStats = (roadmap: SavedRoadmap) => {
-    const lastEdited = roadmap?.timestamp ? new Date(roadmap.timestamp).toLocaleString() : undefined;
-    const roadmapCount = roadmap.planners.length;
-    const courseCount = roadmap.planners
-      .flatMap((planner) => planner.content)
-      .flatMap((year) => year.quarters)
-      .reduce((total, quarter) => total + quarter.courses.length, 0);
-
-    return { lastEdited, roadmapCount, courseCount };
-  };
-
-  // stats for synced roadmap
-  const {
-    lastEdited: syncedLastEdited,
-    roadmapCount: syncedRoadmapCount,
-    courseCount: syncedCourseCount,
-  } = initialAccountRoadmap ? countRoadmapStats(initialAccountRoadmap) : { roadmapCount: 0, courseCount: 0 };
-
-  const {
-    lastEdited: localLastEdited,
-    roadmapCount: localRoadmapCount,
-    courseCount: localCourseCount,
-  } = initialLocalRoadmap ? countRoadmapStats(initialLocalRoadmap) : { roadmapCount: 0, courseCount: 0 };
-
   return (
-    <Dialog
+    <PlannerLoaderModal
       open={showSyncModal}
-      onClose={() => {
-        setShowSyncModal(false);
-      }}
-    >
-      <DialogTitle>Roadmap Out of Sync</DialogTitle>
-      <DialogContent>
-        <DialogContentText>
-          This device's saved roadmap has newer changes than the one saved to your account. Where would you like to load
-          your roadmap from?
-        </DialogContentText>
-        <Divider sx={{ my: 4 }} />
-        {/* Displayed Info */}
-        <div
-          className="displayed-info"
-          style={{
-            display: 'grid',
-            gridTemplateColumns: '1fr 1fr',
-            gap: '16px',
-            marginTop: '12px',
-            marginBottom: '12px',
-          }}
-        >
-          {/* 'This Device' column */}
-          <div style={{ display: 'flex', flexDirection: 'column' }}>
-            <DialogContentText>
-              <strong>This Device</strong>
-            </DialogContentText>
-            <DialogContentText>
-              Last edited: <strong>{getRelativeTime(localLastEdited)}</strong>
-            </DialogContentText>
-            <DialogContentText>
-              <strong>{localRoadmapCount}</strong> roadmaps
-            </DialogContentText>
-            <DialogContentText>
-              <strong>{localCourseCount}</strong> courses
-            </DialogContentText>
-            <Button
-              loading={overrideLoading}
-              disabled={overrideLoading || accountLoading}
-              color="error"
-              variant="contained"
-              onClick={overrideAccountRoadmap}
-              sx={{ marginTop: 4, alignSelf: 'flex-start' }}
-            >
-              This Device
-            </Button>
-            <DialogContentText sx={{ marginTop: 2, fontSize: 12, color: 'red' }}>
-              Warning: Loading from this device can override your synced account data.
-            </DialogContentText>
-          </div>
-          {/* 'My Account' column */}
-          <div style={{ display: 'flex', flexDirection: 'column' }}>
-            <DialogContentText>
-              <strong>My Account</strong>
-            </DialogContentText>
-            <DialogContentText>
-              Last edited: <strong>{getRelativeTime(syncedLastEdited)}</strong>
-            </DialogContentText>
-            <DialogContentText>
-              <strong>{syncedRoadmapCount}</strong> roadmaps
-            </DialogContentText>
-            <DialogContentText>
-              <strong>{syncedCourseCount}</strong> courses
-            </DialogContentText>
-            <Button
-              loading={accountLoading}
-              disabled={overrideLoading || accountLoading}
-              variant="contained"
-              onClick={syncAccount}
-              sx={{ marginTop: 4, alignSelf: 'flex-start' }}
-            >
-              My Account
-            </Button>
-          </div>
-        </div>
-      </DialogContent>
-    </Dialog>
+      onClose={setShowSyncModal}
+      overrideLoading={overrideLoading}
+      accountLoading={accountLoading}
+      initialAccountRoadmap={initialAccountRoadmap}
+      initialLocalRoadmap={initialLocalRoadmap}
+      overrideAccountRoadmap={overrideAccountRoadmap}
+      syncAccount={syncAccount}
+    />
   );
 };
